+++ packages/plugin-rsc/examples/starter/src/root.tsx
@@ -2,7 +2,8 @@ import './index.css' // css import is automatically injected in exported server 
 import viteLogo from '/vite.svg'
 import { getServerCounter, updateServerCounter } from './action.tsx'
 import reactLogo from './assets/react.svg'
-import { ClientCounter, ErrorBoundary } from './client.tsx'
+import { ClientCounter } from './client.tsx'
+import { ErrorBoundary } from './error-boundary.tsx'
 
 export function Root(props: { url: URL }) {
   return (
